packages/embed-api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "embed-api"
version = "0.1.0"
description = "FastAPI service for vector similarity search"
requires-python = ">=3.11"
dependencies = [
    "embed-core",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "pydantic-settings>=2.3",
    "structlog>=24.1",
    "boto3>=1.34",   # for Bedrock provider option
    "Pillow>=10.3",  # for image query endpoint
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/embed_api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
